Add files via upload
After Brexit in Jan 2020, many types of trading were affected. New trading agreements and policies were posted. There are many controversies based on it since the European Union has a prime position when it comes to global trade. Also, the pandemic in 2019 has made significant changes in people's lives. However, challenges also represent opportunities for change and growth. The movement of people between country and country was dramatically reduced, but the opposite is true for goods.  Understanding the trend and historical data can help us understand the logistics industry and identify our important trading partners and opportunities. However, the information is too large and piecemeal; therefore, dashboards or applications are needed to make data more consolidated and easier to understand. This dashboard is not only able to provide insight to logistics companies but also to business that wants to join international trade.